# docs/openapi.yaml
openapi: 3.0.3
info:
  title: deckard API
  version: 1.0.0
  description: |
    REST + SSE API of deckard, the attack-surface monitor.

    Authentication depends on `auth.mode`:
    * `none` - no authentication; read AND write are open. Only run behind a trusted authenticating proxy.
    * `token` - `Authorization: Bearer <token>` on every request except `/healthz` and `/readyz`. The token is never accepted in the query string.
    * `oidc` - browser login via `/auth/login` (Authorization Code + PKCE) and a signed, encrypted session cookie.
      State-changing requests authenticated by cookie must send the `X-CSRF-Token` header with the value returned by `GET /api/v1/me`.

    List responses are `{items,total,limit,offset}`; `limit` defaults to 50 (max 500). Unknown or malformed query parameters yield 400.
    Errors are `{"error":{"code","message"}}`.
servers:
  - url: /
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /auth/login:
    get:
      summary: Start OIDC login (oidc mode only)
      security: []
      parameters:
        - name: next
          in: query
          description: Same-site path to return to after login
          schema: {type: string}
      responses:
        '302': {description: Redirect to the identity provider}
        '502': {$ref: '#/components/responses/Error'}
  /auth/callback:
    get:
      summary: OIDC redirect URI (oidc mode only)
      security: []
      parameters:
        - {name: code, in: query, schema: {type: string}}
        - {name: state, in: query, schema: {type: string}}
        - {name: error, in: query, schema: {type: string}}
      responses:
        '302': {description: Session established; redirect to the original page}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '502': {$ref: '#/components/responses/Error'}
  /auth/logout:
    post:
      summary: End the session (oidc mode only)
      description: |
        POST only (a GET would be triggerable cross-site). With a live session the
        `X-CSRF-Token` header (from `GET /api/v1/me` `csrf_token`) is required.
      security: []
      parameters:
        - {name: X-CSRF-Token, in: header, schema: {type: string}}
      responses:
        '303': {description: Cookie cleared; redirect to /}
        '403': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
    sessionCookie: {type: apiKey, in: cookie, name: __Host-deckard_session, description: "Unprefixed deckard_session when base_url is not https"}
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
...